Implemented Action Cards
- To create action cards for a card:
  1. go under the scriptable object CardNode of that card
  2. add new element under the Skills dropdown
  3. MUST fill in sprite for Card Sprite, and all fields of Effect 1.

*TODO:
1. Implement Effect 2 values to allow skills like "Attack an enemy and heal an ally"
2. Implement attack program to allow multiple attacks/multiple targets with one attack.

+++ Assets/Scripts/Combat/Fight Node.cs
@@ -8,7 +8,7 @@ using UnityEngine.UI;
 public class FightNode : ScriptableObject
 {
     public int enemyName;
-    public float enemyHP;
+    public int enemyHP;
     public int enemyATK;
     public int enemySPD;
